% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{create_palette}
\alias{create_palette}
\alias{is_palette}
\title{\code{palette} vector}
\usage{
create_palette(pal, n = length(pal), type = c("discrete", "continuous"))

is_palette(pal)
}
\arguments{
\item{pal}{\itemize{
\item For \code{create_palette()}: A character vector of hex codes
\item For \code{is_palette()}: An object to test
}}

\item{n}{The number of colors}

\item{type}{The type of palette, either \code{discrete} or \code{continuous}}
}
\description{
This creates a character vector that represents palettes so when it is
printed, it displays the palette colors.
}
\examples{
create_palette(pal = c("#842000", "#EC9B01", "#3F5D91"), n = 3L)
}
