% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-access.R
\name{get_spotify_access_token}
\alias{get_spotify_access_token}
\title{Spotify API helpers}
\usage{
get_spotify_access_token()
}
\value{
The Spotify access token.
}
\description{
This is a wrapper around \code{\link[spotifyr:get_spotify_access_token]{spotifyr::get_spotify_access_token()}} to create a
Spotify access token from stored environment variables.
}
\examples{
\dontshow{if (taylor_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{get_spotify_access_token()}
\dontshow{\}) # examplesIf}
}
