% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_series.R
\name{make_taylor_series}
\alias{make_taylor_series}
\title{Taylor Series Converter}
\usage{
make_taylor_series(
  f,
  num_derivs,
  center,
  intr_start,
  intr_end,
  var_name = "x",
  open_closed = "closed"
)
}
\arguments{
\item{f}{The function you want to convert into a Taylor Series}

\item{num_derivs}{How derivatives you want to take of f}

\item{center}{The center to test. In Stewart Calculus, this is notated "a"}

\item{intr_start}{The lower bound to try for the remainder interval}

\item{intr_end}{The upper bound to try for the remainder interval}

\item{var_name}{Differentiating f w.r.t var_name. Can only be one variable name
at the moment. The default variable name is "x".}

\item{open_closed}{Whether or not the interval defined by sequence
from intr_start to intr_end}
}
\description{
The Taylor Series Converter Function performs a single variable Taylor
Expansion of the given function. It returns the call in full, the error
approximation, and the Taylor Remainder graph.
}
