% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_series.R
\name{R_nx}
\alias{R_nx}
\title{Taylor Remainder}
\usage{
R_nx(df_np1, x, n, a)
}
\arguments{
\item{df_np1}{The n+1th derivative of the last derivative present in the
Taylor polynomial}

\item{x}{the value of x to plug into the absolute value of the
n+1th derivative}

\item{n}{The integer number of derivatives taken in total}

\item{a}{The center being tested}
}
\value{
Returns the absolute value of n+1th derivative multiplied by
x - a raised to a power of n+1 all divided by the factorial of n+1
}
\description{
The Taylor Remainder Function is integrated with the make_taylor_series
function to provide an error estimate when calculating the Taylor series
for a given function.
}
