% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-color-scales.R
\name{scale_fill_albums}
\alias{scale_fill_albums}
\alias{scale_colour_albums}
\alias{scale_color_albums}
\title{Taylor Swift album comparison palette}
\usage{
scale_fill_albums(
  ...,
  aesthetics = "fill",
  breaks = waiver(),
  limits = force,
  na.value = "grey50"
)

scale_colour_albums(...)

scale_color_albums(...)
}
\arguments{
\item{...}{Arguments to be passed to}

\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale values
\item A character vector that defines possible values of the scale and their
order
\item A function that accepts the existing (automatic) values and returns
new ones. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{na.value}{If \code{na.translate = TRUE}, what aesthetic value should the
missing values be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}
}
\description{
something
}
