% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-access.R
\name{soundstat-api}
\alias{soundstat-api}
\alias{get_soundstat_api_key}
\alias{set_soundstat_api_key}
\title{SoundStat API helpers}
\usage{
get_soundstat_api_key()

set_soundstat_api_key(key = NULL)
}
\arguments{
\item{key}{A SoundStat API key.}
}
\value{
\itemize{
\item \code{get_soundstat_api_key()} returns a previously stored API key.
\item \code{set_soundstat_api_key()} is called for side effects only.
}
}
\description{
Set and retrieve a SoundStat API key
}
\examples{
\dontshow{if (taylor_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{get_soundstat_api_key()}
\dontshow{\}) # examplesIf}
}
