% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-access.R
\name{spotify-api}
\alias{spotify-api}
\alias{get_spotify_access_token}
\alias{set_spotify_api_key}
\title{Spotify API helpers}
\usage{
get_spotify_access_token()

set_spotify_api_key(id = NULL, secret = NULL)
}
\arguments{
\item{id}{A Spotify Client ID.}

\item{secret}{A Spotify Client Secret.}
}
\value{
\itemize{
\item \code{get_spotify_api_key()} returns a previously stored Client ID and Secret.
\item \code{set_spotify_api_key()} is called for side effects only.
}
}
\description{
Set and retrieve Spotify client information.
}
\examples{
\dontshow{if (taylor_examples()) withAutoprint(\{ # examplesIf}
\donttest{get_spotify_access_token()}
\dontshow{\}) # examplesIf}
}
