% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor-package.R
\docType{package}
\name{taylor}
\alias{taylor}
\title{taylor: Lyrics and song data for Taylor Swift's discography}
\description{
A comprehensive resource for data on Taylor Swift songs. Data for all
officially released studio albums, extended plays (EPs), and individual
singles are included. Data comes from \url{https://genius.com/} (lyrics) and
\url{https://www.spotify.com/us/} (song characteristics).
}
