% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-access.R
\name{taylor_examples}
\alias{taylor_examples}
\title{Determine if code is executed interactively or in pkgdown}
\usage{
taylor_examples()
}
\value{
A logical value indicating whether or not the examples should be run.
}
\description{
Used for determining examples that shouldn't be run on CRAN, but can be run
for the pkgdown website.
}
\examples{
taylor_examples()
}
