% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{taylor_sitrep}
\alias{taylor_sitrep}
\title{Situation report for APIs}
\usage{
taylor_sitrep()
}
\value{
Called for side effects.
}
\description{
Check configuration status for Spotify and SoundStat APIs.
}
\examples{
\dontshow{if (taylor_examples()) withAutoprint(\{ # examplesIf}
taylor_sitrep()
\dontshow{\}) # examplesIf}
}
