% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{title_case}
\alias{title_case}
\title{Convert string to title case}
\usage{
title_case(string)
}
\arguments{
\item{string}{String to modify.}
}
\value{
A character string with the first letter of each word capitalized.
}
\description{
Capitalize the first letter of each word, and convert the remaining string to
lower case.
}
\examples{
title_case("taylor swift")
title_case("Taylor Swift")
title_case("TAYLOR SWIFT")
}
