% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyric-abbreviation.R
\name{translate_bracelet}
\alias{translate_bracelet}
\title{Translate a friendship bracelet}
\usage{
translate_bracelet(abbr)
}
\arguments{
\item{abbr}{The abbreviated lyrics (i.e., the first letter of each word).}
}
\value{
A character vector with the name of the song and the abbreviated
line.
}
\description{
Receive a friendship bracelet at the Eras Tour but can't figure out what
lyrics the bracelet has abbreviated? Now you can find out!
}
\examples{
translate_bracelet("bykmosftvfw")

translate_bracelet("kimbkiagkitbimhotw")
}
