% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tswift_album_songs}
\alias{tswift_album_songs}
\title{Data for Songs on Taylor's Official Studio Albums}
\format{
This is a subset of the \code{\link{tswift_all_songs}} data set. Please see
that documentation for a complete description of all the included fields.
}
\usage{
tswift_album_songs
}
\description{
A data set containing lyrics to and characteristics of songs on all of
Taylor's official studio albums. Thus, this is a subset of
\code{\link{tswift_all_songs}}, with EPs and individual singles excluded. Critically,
this subset also only includes versions owned by Taylor when possible,
because we stan artists owning their own work. This means that although both
\emph{Fearless} and \emph{Fearless (Taylor's Version)} appear in full
\code{\link{tswift_all_songs}} data, only \emph{Fearless (Taylor's Version)} appears in this
subset. This also means that this data set will change as additional
re-releases are made available (i.e., \emph{Red} will soon be replaced with
\emph{Red (Taylor's Version)}).
}
\keyword{datasets}
