% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor-album-palettes.R
\docType{data}
\name{album_palettes}
\alias{album_palettes}
\title{Color palettes}
\format{
An object of class \code{list} of length 9.
}
\usage{
album_palettes
}
\description{
Colors derived from album covers at \url{http://colormind.io/image/}.
}
\keyword{datasets}
