% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor-package.R
\docType{package}
\name{taylor-package}
\alias{taylor}
\alias{taylor-package}
\title{taylor: Lyrics and Song Data for Taylor Swift's Discography}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive resource for data on Taylor Swift songs. Data is included for all officially released studio albums, extended plays (EPs), and individual singles are included. Data comes from 'Genius' (lyrics) and 'Spotify' (song characteristics). Additional functions are included for easily creating data visualizations with color palettes inspired by Taylor Swift's album covers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://taylor.wjakethompson.com}
  \item \url{https://github.com/wjakethompson/taylor}
  \item Report bugs at \url{https://github.com/wjakethompson/taylor/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID}) [copyright holder]

}
\keyword{internal}
